requirements.txt
streamlit
PyMuPDF
python-docx
nltk
